+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: anlima <anlima@student.42lisboa.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/13 15:55:51 by anlima            #+#    #+#             */
-/*   Updated: 2022/08/13 16:30:24 by anlima           ###   ########.fr       */
+/*   Updated: 2022/10/04 10:22:06 by anlima           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	char	*str1;
 	char	*str2;
-	int	i;
+	size_t	i;
 
 	str1 = (char *)s1;
 	str2 = (char *)s2;
