+++ ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: anlima <anlima@student.42lisboa.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/13 12:57:10 by anlima            #+#    #+#             */
-/*   Updated: 2022/08/13 13:13:44 by anlima           ###   ########.fr       */
+/*   Updated: 2022/10/04 10:14:40 by anlima           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void *ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 	char *str;
 
 	str = (char *)s;
